# src/main/resources/api/dev-practice.openapi.yaml
openapi: 3.0.3
info:
  title: La Forge API
  version: 0.1.0
  description: |
    Contrat du prototype de laboratoire d’entraînement technique. Texte métier en français, identifiants API en anglais.

    Périmètre : profil, catalogue, exercices versionnés, tentatives, indices, corrigés, révisions, veille, défis collectifs, édition et tuteur IA optionnel. Pas d’exécution distante de code, génération automatique de contenus, classement RH ou gestion multi-entreprises. Une installation correspond à une entreprise ou un groupe pilote. Les rôles sont gérés hors API.

    Authentification : compte créé localement via /auth/registrations (email + mot de passe), activé par confirmation
    du jeton de vérification reçu par courriel (/auth/verifications). Connexion via /auth/sessions : le serveur émet
    lui-même un bearer JWT de courte durée (signature RS256) accompagné d’un jeton de renouvellement opaque
    (/auth/sessions/refresh). Le serveur valide signature, issuer, audience et expiration ; utilisateur courant déduit
    de sub, jamais du corps de requête. Rôles LEARNER, AUTHOR, REVIEWER, ADMIN cumulatifs ; tous les comptes peuvent
    pratiquer dès activation. x-required-roles signifie au moins un rôle de la liste, plus les contraintes de propriété
    décrites. ADMIN n’accorde pas l’accès aux tentatives privées des autres.

...

...

...

...

...

...

...
servers:
- url: /api/v1
  description: Même origine que le frontend ; préfixe de la v1.
security:
- bearerAuth: []
...
paths:
...
  /authoring/drafts:
    get:
      tags:
      - Authoring
      operationId: listDrafts
      summary: Lister les brouillons accessibles
      description: 'AUTHOR : ses brouillons ; REVIEWER : brouillons IN_REVIEW et ceux qu’il a déjà relus ; ADMIN
        : tous. Rôles cumulatifs. Ordre updatedAt décroissant puis id. Contenu incluant le corrigé réservé à ce
        périmètre éditorial.'
      responses:
        '200':
          description: Lister les brouillons accessibles.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - name: state
        in: query
        schema:
          type: string
          enum:
          - DRAFT
          - IN_REVIEW
          - APPROVED
          - PUBLISHED
        description: ''
      x-required-roles:
      - AUTHOR
      - REVIEWER
      - ADMIN
    post:
      tags:
      - Authoring
      operationId: createDraft
      summary: Créer un brouillon ou une nouvelle version
      description: 'AUTHOR ou ADMIN. Nouvel exercice : exerciseId attribué côté serveur et baseVersion=0. Révision
        existante : propriétaire de l’exercice ou ADMIN, baseVersion figée sur dernière publication. Un seul brouillon
        actif par exercice, sinon 409. État DRAFT, revision=1.'
      responses:
        '201':
          description: Créer un brouillon ou une nouvelle version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
          headers:
            ETag:
              description: Version forte du brouillon, par exemple "7". À recopier dans If-Match.
              schema:
                type: string
              example: '"7"'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftInput'
      x-required-roles:
      - AUTHOR
      - ADMIN
  /authoring/drafts/{draftId}:
    get:
      tags:
      - Authoring
      operationId: getDraft
      summary: Lire un brouillon et son ETag
      description: Même périmètre que listDrafts ; sinon 404.
      responses:
        '200':
          description: Lire un brouillon et son ETag.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
          headers:
            ETag:
              description: Version forte du brouillon, par exemple "7". À recopier dans If-Match.
              schema:
                type: string
              example: '"7"'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: draftId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      x-required-roles:
      - AUTHOR
      - REVIEWER
      - ADMIN
    put:
      tags:
      - Authoring
      operationId: replaceDraft
      summary: Remplacer le contenu du brouillon
      description: Auteur propriétaire ou ADMIN ; DRAFT uniquement, sinon 409. Remplacement atomique, incrémente
        revision. Les anciennes reviews restent dans l’historique mais ne valent jamais approbation de la nouvelle
        révision.
      responses:
        '200':
          description: Remplacer le contenu du brouillon.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
          headers:
            ETag:
              description: Version forte du brouillon, par exemple "7". À recopier dans If-Match.
              schema:
                type: string
              example: '"7"'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: draftId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftUpdate'
      x-required-roles:
      - AUTHOR
      - ADMIN
  /authoring/drafts/{draftId}/submit:
    post:
      tags:
      - Authoring
      operationId: submitDraftForReview
      summary: Soumettre le brouillon à relecture
      description: Auteur propriétaire ou ADMIN. DRAFT → IN_REVIEW. Vérifie cohérence complète des choix, corrigé,
        indices, références et thèmes. Incrémente revision. Aucune publication automatique.
      responses:
        '200':
          description: Soumettre le brouillon à relecture.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
          headers:
            ETag:
              description: Version forte du brouillon, par exemple "7". À recopier dans If-Match.
              schema:
                type: string
              example: '"7"'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: draftId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IfMatch'
      x-required-roles:
      - AUTHOR
      - ADMIN
  /authoring/drafts/{draftId}/review:
    post:
      tags:
      - Authoring
      operationId: reviewDraft
      summary: Approuver ou demander des changements
      description: REVIEWER ou ADMIN, IN_REVIEW uniquement. APPROVE → APPROVED ; REQUEST_CHANGES → DRAFT. Incrémente
        revision et enregistre la décision. Pour un pilote individuel, autoapprobation autorisée uniquement à ADMIN
        et consignée dans reviews ; REVIEWER ne peut pas approuver son propre contenu (403).
      responses:
        '200':
          description: Approuver ou demander des changements.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
          headers:
            ETag:
              description: Version forte du brouillon, par exemple "7". À recopier dans If-Match.
              schema:
                type: string
              example: '"7"'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: draftId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewInput'
      x-required-roles:
      - REVIEWER
      - ADMIN
  /authoring/drafts/{draftId}/publish:
    post:
      tags:
      - Authoring
      operationId: publishDraft
      summary: Publier la version approuvée
      description: 'Auteur propriétaire ou ADMIN. APPROVED requis. Vérifie que baseVersion est toujours la dernière
        version publiée, sinon 409. Publication atomique : nouvelle version immuable baseVersion+1, brouillon PUBLISHED
        et revision incrémentée. Corrigé et indices stockés avec cette version ; seules les projections apprenant
        sont retournées. Nouvelle modification : nouveau brouillon.'
      responses:
        '200':
          description: Publier la version approuvée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: draftId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IfMatch'
      x-required-roles:
      - AUTHOR
      - ADMIN
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Jeton du fournisseur d’identité configuré. Acquisition hors périmètre.
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Nombre maximal d’éléments.
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 1000
      description: Curseur opaque retourné par la page précédente.
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        maxLength: 100
        minLength: 1
      description: 'ETag fort exact obtenu sur le brouillon. Pas de joker. Absent : 428 ; obsolète : 412.'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: 'UUID généré par le client, unique pour utilisateur + méthode + chemin. Conservation 24 h. Même
        clé et même corps : réponse originale sans répéter l’effet. Autre corps : 409. Déduplication atomique, y
        compris des appels IA. Opération encore en cours : 409 INVALID_STATE, réessayer avec la même clé.'
  responses:
    BadRequest:
      description: Requête syntaxiquement invalide.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthenticated:
      description: Jeton absent, expiré ou invalide.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        WWW-Authenticate:
          description: Challenge Bearer.
          schema:
            type: string
          example: Bearer
    Forbidden:
      description: Rôle insuffisant ou déblocage pédagogique non acquis.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Ressource inexistante ou non accessible dans le périmètre de l’appelant.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Transition interdite, tentative déjà finalisée, version de base obsolète ou conflit d’idempotence.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    PreconditionFailed:
      description: ETag du brouillon obsolète.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationFailed:
      description: Données ou contraintes métier invalides ; violations détaille les champs.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    PreconditionRequired:
      description: If-Match requis.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    TooManyRequests:
      description: Quota dépassé ; respecter Retry-After.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        Retry-After:
          description: Délai minimal en secondes avant nouvel essai, si disponible.
          schema:
            type: integer
            format: int32
            minimum: 0
    InternalError:
      description: Erreur serveur inattendue.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    ExerciseType:
      type: string
      enum:
      - CODE_REVIEW
      - PREDICTION
      - DIAGNOSIS
      - REFACTORING
      - TECH_DISCOVERY
      - TECHNICAL_CHOICE
      - QUIZ
    Difficulty:
      type: string
      enum:
      - BEGINNER
      - INTERMEDIATE
      - ADVANCED
    Problem:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
          format: uri-reference
          example: urn:dev-practice:problem:invalid-state
        title:
          type: string
          maxLength: 200
          minLength: 1
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
        detail:
          type: string
          maxLength: 4000
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          enum:
          - BAD_REQUEST
          - UNAUTHENTICATED
          - FORBIDDEN
          - NOT_FOUND
          - INVALID_STATE
          - IDEMPOTENCY_CONFLICT
          - STALE_VERSION
          - VALIDATION_FAILED
          - PRECONDITION_REQUIRED
          - RATE_LIMITED
          - AI_UNAVAILABLE
          - INTERNAL_ERROR
        traceId:
          type: string
          maxLength: 100
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
          maxItems: 100
      required:
      - type
      - title
      - status
      - code
      example:
        type: urn:dev-practice:problem:invalid-state
        title: Tentative déjà soumise
        status: 409
        code: INVALID_STATE
        detail: La réponse définitive ne peut plus être modifiée.
    Violation:
      type: object
      additionalProperties: false
      properties:
        field:
          type: string
          maxLength: 200
          minLength: 1
        message:
          type: string
          maxLength: 4000
      required:
      - field
      - message
...
    TechnologyRequirement:
      type: object
      additionalProperties: false
      properties:
        technology:
          type: string
          maxLength: 200
          minLength: 1
        minimumVersion:
          type: string
          maxLength: 40
          minLength: 1
        featureStatus:
          type: string
          enum:
          - STABLE
          - PREVIEW
          - INCUBATOR
        notes:
          type: string
          maxLength: 4000
      required:
      - technology
      - minimumVersion
      - featureStatus
    CodeFile:
      type: object
      additionalProperties: false
      properties:
        path:
          type: string
          maxLength: 240
          minLength: 1
          description: Chemin relatif logique, jamais utilisé directement comme chemin disque.
        language:
          type: string
          maxLength: 40
          minLength: 1
          example: java
        content:
          type: string
          maxLength: 50000
        kind:
          type: string
          enum:
          - SOURCE
          - DIFF
          - LOG
          - TEST
      required:
      - path
      - language
      - content
      - kind
    Choice:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          maxLength: 50
          pattern: ^[a-zA-Z0-9_-]+$
        label:
          type: string
          maxLength: 2000
          minLength: 1
      required:
      - id
      - label
    ResponseSpec:
      type: object
      additionalProperties: false
      properties:
        kind:
          type: string
          enum:
          - FREE_TEXT
          - SINGLE_CHOICE
          - MULTIPLE_CHOICE
          - REVIEW
        choices:
          type: array
          items:
            $ref: '#/components/schemas/Choice'
          maxItems: 10
      required:
      - kind
      description: SINGLE_CHOICE/MULTIPLE_CHOICE exigent 2 à 10 choix aux identifiants uniques. FREE_TEXT/REVIEW
        interdisent choices. QUIZ exige un format de choix. Les autres types peuvent utiliser tout format compatible
        avec leur énoncé. Validation serveur, erreur 422.
...
    Exercise:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          format: int32
          minimum: 1
        title:
          type: string
          maxLength: 200
          minLength: 1
        type:
          $ref: '#/components/schemas/ExerciseType'
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        topicIds:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 20
          uniqueItems: true
        estimatedMinutes:
          type: integer
          format: int32
          minimum: 1
          maximum: 60
        technologies:
          type: array
          items:
            $ref: '#/components/schemas/TechnologyRequirement'
          maxItems: 10
        publishedAt:
          type: string
          format: date-time
        promptMarkdown:
          type: string
          maxLength: 20000
          minLength: 1
        learningObjectives:
          type: array
          items:
            type: string
            maxLength: 200
            minLength: 1
          maxItems: 10
          minItems: 1
        files:
          type: array
          items:
            $ref: '#/components/schemas/CodeFile'
          maxItems: 10
        responseSpec:
          $ref: '#/components/schemas/ResponseSpec'
        hintCount:
          type: integer
          format: int32
          minimum: 0
          maximum: 5
      required:
      - id
      - version
      - title
      - type
      - difficulty
      - topicIds
      - estimatedMinutes
      - technologies
      - publishedAt
      - promptMarkdown
      - learningObjectives
      - files
      - responseSpec
      - hintCount
      description: Projection apprenant. Ne contient ni corrigé, ni bonne réponse, ni contenu des indices non débloqués.
        Une version publiée est immuable.
    RubricCriterion:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          maxLength: 50
          minLength: 1
        label:
          type: string
          maxLength: 200
          minLength: 1
        importance:
          type: string
          enum:
          - BLOCKING
          - MAJOR
          - MINOR
        explanationMarkdown:
          type: string
          maxLength: 20000
          minLength: 1
      required:
      - id
      - label
      - importance
      - explanationMarkdown
    Correction:
      type: object
      additionalProperties: false
      properties:
        explanationMarkdown:
          type: string
          maxLength: 20000
          minLength: 1
        acceptedAnswers:
          type: array
          items:
            type: string
            maxLength: 4000
            minLength: 1
          maxItems: 20
        correctChoiceIds:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 10
          uniqueItems: true
        expectedVerdicts:
          type: array
          items:
            type: string
            enum:
            - APPROVE
            - REQUEST_CHANGES
            - NEED_INFORMATION
          maxItems: 3
          uniqueItems: true
        criteria:
          type: array
          items:
            $ref: '#/components/schemas/RubricCriterion'
          maxItems: 20
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
          maxItems: 20
          minItems: 1
      required:
      - explanationMarkdown
      - acceptedAnswers
      - criteria
      - sources
      description: Pour un choix, correctChoiceIds est obligatoire et référence les choix proposés (exactement un
        pour SINGLE_CHOICE). Pour REVIEW, expectedVerdicts est obligatoire et non vide. Les propriétés incompatibles
        sont interdites. Une réponse libre ne reçoit pas de note automatique de qualité.
    Hint:
      type: object
      additionalProperties: false
      properties:
        level:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
        markdown:
          type: string
          maxLength: 20000
          minLength: 1
      required:
      - level
      - markdown
    ExerciseContent:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
        type:
          $ref: '#/components/schemas/ExerciseType'
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        topicIds:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 20
          uniqueItems: true
        estimatedMinutes:
          type: integer
          format: int32
          minimum: 1
          maximum: 60
        technologies:
          type: array
          items:
            $ref: '#/components/schemas/TechnologyRequirement'
          maxItems: 10
        promptMarkdown:
          type: string
          maxLength: 20000
          minLength: 1
        learningObjectives:
          type: array
          items:
            type: string
            maxLength: 200
            minLength: 1
          maxItems: 10
          minItems: 1
        files:
          type: array
          items:
            $ref: '#/components/schemas/CodeFile'
          maxItems: 10
        responseSpec:
          $ref: '#/components/schemas/ResponseSpec'
        hints:
          type: array
          items:
            $ref: '#/components/schemas/Hint'
          maxItems: 5
        correction:
          $ref: '#/components/schemas/Correction'
      required:
      - title
      - type
      - difficulty
      - topicIds
      - estimatedMinutes
      - technologies
      - promptMarkdown
      - learningObjectives
      - files
      - responseSpec
      - hints
      - correction
      description: Contenu éditorial complet, réservé à auteur habilité et relecteur. Indices contigus à partir
        de 1. Les références, versions, choix et critères sont vérifiés avant publication.
    DraftInput:
      type: object
      additionalProperties: false
      properties:
        exerciseId:
          type: string
          format: uuid
        content:
          $ref: '#/components/schemas/ExerciseContent'
      required:
      - content
      description: 'exerciseId absent : nouvel exercice. Présent : nouvelle version d’un exercice dont l’appelant
        est auteur ou administrateur. Une seule révision non publiée par exercice. Le serveur mémorise la version
        publiée de base.'
    DraftUpdate:
      type: object
      additionalProperties: false
      properties:
        content:
          $ref: '#/components/schemas/ExerciseContent'
      required:
      - content
    EditorialReview:
      type: object
      additionalProperties: false
      properties:
        reviewerId:
          type: string
          format: uuid
        decision:
          type: string
          enum:
          - APPROVE
          - REQUEST_CHANGES
        comment:
          type: string
          maxLength: 20000
          minLength: 1
        reviewedAt:
          type: string
          format: date-time
      required:
      - reviewerId
      - decision
      - comment
      - reviewedAt
    Draft:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        exerciseId:
          type: string
          format: uuid
        authorId:
          type: string
          format: uuid
        state:
          type: string
          enum:
          - DRAFT
          - IN_REVIEW
          - APPROVED
          - PUBLISHED
        revision:
          type: integer
          format: int32
          minimum: 1
        baseVersion:
          type: integer
          format: int32
          minimum: 0
        content:
          $ref: '#/components/schemas/ExerciseContent'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/EditorialReview'
          maxItems: 100
        publishedVersion:
          type: integer
          format: int32
          minimum: 1
        updatedAt:
          type: string
          format: date-time
      required:
      - id
      - exerciseId
      - authorId
      - state
      - revision
      - baseVersion
      - content
      - reviews
      - updatedAt
    ReviewInput:
      type: object
      additionalProperties: false
      properties:
        decision:
          type: string
          enum:
          - APPROVE
          - REQUEST_CHANGES
        comment:
          type: string
          maxLength: 20000
          minLength: 1
      required:
      - decision
      - comment
...
    DraftPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Draft'
          maxItems: 100
        nextCursor:
          type: string
          maxLength: 1000
      required:
      - items
      description: Pagination par curseur opaque ; nextCursor absent à la dernière page. Aucun total garanti. Un
        curseur doit être réutilisé avec les mêmes filtres.
...